Added observer deployment trend code.
The program "observer_turnout_trends.py" looks at how the relative (i.e.
subtracting off the median) percent change in observer deployment
density correlates with the percent change in turnout between the
first-round and runoff elections. Various graphs related to this have
been added to the "clean_data" folder, along with a CSV that details the
mapping between province numbers and province names.

I've also started off a script ("afghan_constants.py") that contains
various constants for this election, such as the voting-eligible
fraction of the population (which is close to 66%).