Moved first round votes data to clean_data.
Now, the raw_data folder contains the initial data as downloaded from
AfghanistanElectionData.org, while the clean_data folder contains the
vote data that I cleaned (so that the district names match up with the
runoff results, etc). The raw_data CSV file is just included for
reference, and should not actually be used.